% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{browse_tables}
\alias{browse_tables}
\title{Show index of tables from the AOFM website}
\usage{
browse_tables(x = aofm_index_nav)
}
\arguments{
\item{x}{object you want data on, options include: tb, tib, tn, slf, summary, aggregate, ownership, retail, termpremium}
}
\value{
dataframe with index of files and arguments to download them
}
\description{
Show index of tables from the AOFM website
}
\examples{
\dontrun{download_aofm_xlsx()}
# return dataframe with list of files/arguments
}
