% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_excel.R
\name{download_aofm_xlsx}
\alias{download_aofm_xlsx}
\title{Download xlsx.xls files from AOFM DataHub}
\usage{
download_aofm_xlsx(security = NULL, type = NULL, data.index = aofm_index)
}
\arguments{
\item{security}{object you want data on, options include:
tb, tib, tn, slf, summary, aggregate, ownership, retail, termpremium}

\item{type}{the specific type of data you want, occasionally optional:
dealt, settlement, issuance, syndication, buyback, turnover,}

\item{data.index}{internal data required to look up file to download}
}
\value{
a list of downloaded files
}
\description{
Download xlsx.xls files from AOFM DataHub
}
\examples{
\dontrun{download_aofm_xlsx("tb", "issuance")}
# downloads AOFM Treasury Bond Issuance data to data/tb_issuance.xlsx
\dontrun{download_aofm_xlsx("tb")}
# downloads all Treasury Bond data from the AOFM Datahub incl
# issunace, syndication details, position data etc to data/
\dontrun{download_aofm_xlsx()}
# downloads all data from the AOFM Datahub to data/

}
