% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_file}
\alias{find_file}
\title{Find name of file to download from AOFM}
\usage{
find_file(security = NULL, type = NULL, data.index = aofm_index)
}
\arguments{
\item{security}{object you want data on, options include: tb, tib, tn, slf, summary, aggregate, ownership, retail, termpremium}

\item{type}{the specific type of data you want, occasionally optional: dealt, settlement, issuance, syndication, buyback, turnover,}

\item{data.index}{internal data required to look up file to download}
}
\value{
a vector with the file/s which match the input params
}
\description{
Find name of file to download from AOFM
}
\examples{
find_file("tb", "issuance")
returns "tb_issuance"
find_file("tb")
returns c("tb_issuance", "tb_syndication", "tb_turnover" .....)
find_file()
returns vector of all filenames
}
