% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{not_all_na}
\alias{not_all_na}
\title{Select columns which dont have all NA values, cc zack on StackOverflow}
\usage{
not_all_na(x)
}
\arguments{
\item{datframe}{dataframe object}
}
\value{
columns which are not ONLY NA values
}
\description{
Select columns which dont have all NA values, cc zack on StackOverflow
}
