% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aofm.R
\name{read_aofm}
\alias{read_aofm}
\title{Reads in and cleans any	data from the AOFM, returns cleaned in a long format}
\usage{
read_aofm(security = NULL, type = NULL, csv = F)
}
\arguments{
\item{security}{object you want data on, options include:
tb, tib, tn, slf, summary, aggregate, ownership, retail, termpremium}

\item{type}{the specific type of data you want, occasionally optional:
dealt, settlement, issuance, syndication, buyback, turnover,}

\item{csv}{do you want to export the cleaned data as a csv to /data}

\item{data.index}{internal data required to look up file to download}
}
\value{
AOFM selected AOFM data as a dataframe or list
}
\description{
Reads in and cleans any	data from the AOFM, returns cleaned in a long format
}
\examples{
\dontrun{read("tb", "issuance")}
# downloads Treasury Bond Issunace data from the AOFM
# returns cleaned in a long format

}
