% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_table_functions.R
\name{read_eom}
\alias{read_eom}
\title{Reads in and cleans the "End of Month Positions" data from the AOFM, returns a list}
\usage{
read_eom(aofm_table, csv = F)
}
\arguments{
\item{aofm_table}{object to download, typically called from read_aofm()}

\item{csv}{do you want to export the cleaned data as a csv to /data}
}
\value{
AOFM eom data is returned as a list with data frame elements for
FaceValue, MarketValue, Delta, Duration and Tenor
}
\description{
Reads in and cleans the "End of Month Positions" data from the AOFM, returns a list
}
\examples{
\dontrun{ read_eom("tb_position_dealt")}
# downloads Treasury Bond EOM data on a dealt basis from the AOFM
# returns cleaned in a long format
# data is returned as a list with data frame elements for
# FaceValue, MarketValue, Delta, Duration and Tenor
\dontrun{read_eom("tb_position_settlement")}
# downloads Treasury Bond EOM data on a settlement basis from the
# AOFM, returns cleaned in a long format as a list with elements for each of
# FaceValue, MarketValue, Delta, Duration and Tenor

}
