% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read multiple excel sheet to a list object, cc Ralf Stubner on StackOverflow}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{The file path to the .xlsx from which you want to pull all sheets}
}
\value{
list object with dataframe for each sheet
}
\description{
Read multiple excel sheet to a list object, cc Ralf Stubner on StackOverflow
}
