% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_table_functions.R
\name{read_ownership}
\alias{read_ownership}
\title{Reads in and cleans data under "Ownership of Australia Government Securities" from the AOFM, returns cleaned in a long format}
\usage{
read_ownership(aofm_table, csv = F)
}
\arguments{
\item{aofm_table}{object to download, typically called from read_aofm()}

\item{csv}{do you want to export the cleaned data as a csv to /data}
}
\value{
AOFM ownership data as a list, with elements relating to sheets in the original file
}
\description{
Reads in and cleans data under "Ownership of Australia Government Securities" from the AOFM, returns cleaned in a long format
}
\examples{
\dontrun{read_ownership("ownership_nonresident")}

}
