% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_table_functions.R
\name{read_premium}
\alias{read_premium}
\title{Reads in and cleans data under "Term Premium Estimates" from the AOFM, returns cleaned in a long format}
\usage{
read_premium(aofm_table, csv = F, data.index = aofm_index)
}
\arguments{
\item{aofm_table}{object to download, typically called from read_aofm()}

\item{csv}{do you want to export the cleaned data as a csv to /data}

\item{data.index}{internal data required to look up file to download}
}
\value{
AOFM Transactional data as a dataframe
}
\description{
Reads in and cleans data under "Term Premium Estimates" from the AOFM, returns cleaned in a long format
}
\examples{
read_premium("termpremium")
downloads Term Premium data from the AOFM, returns cleaned in a long format

}
