% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_table_functions.R
\name{read_transactional}
\alias{read_transactional}
\title{Reads in and cleans data under "Transaction Details" from the AOFM, returns cleaned in a long format (excl. issunace via conversion and syndication details)}
\usage{
read_transactional(aofm_table, csv = F, data.index = aofm_index)
}
\arguments{
\item{aofm_table}{object to download, typically called from read_aofm()}

\item{csv}{do you want to export the cleaned data as a csv to /data}

\item{data.index}{internal data required to look up file to download}
}
\value{
AOFM Transactional data as a dataframe
}
\description{
Reads in and cleans data under "Transaction Details" from the AOFM, returns cleaned in a long format (excl. issunace via conversion and syndication details)
}
\examples{
read_transactional("tb_issunace")
downloads Treasury Bond Issunace data from the AOFM, returns cleaned in a long format
read_transactional("slf")
downloads Securities Lending Facility data from the AOFM, returns cleaned in a long format
}
